requirements.txt
discord.py==2.3.1
topggpy==1.4.0
dateutil==2.8.2
aiohttp==3.8.4